+++ server/plus.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 
 	rportplus "github.com/realvnc-labs/rport/plus"
+	"github.com/realvnc-labs/rport/plus/capabilities/extendedpermission"
 	licensecap "github.com/realvnc-labs/rport/plus/capabilities/license"
 	"github.com/realvnc-labs/rport/plus/capabilities/oauth"
-	"github.com/realvnc-labs/rport/plus/capabilities/permission"
 	"github.com/realvnc-labs/rport/plus/capabilities/status"
 	"github.com/realvnc-labs/rport/plus/license"
 	"github.com/realvnc-labs/rport/server/chconfig"
@@ -108,15 +108,15 @@ func RegisterPlusCapabilities(plusManager rportplus.Manager, cfg *chconfig.Confi
 	}
 	logger.Infof("plus license capability registered")
 
-	// always register the plus permission capability
-	_, err = plusManager.RegisterCapability(rportplus.PlusPermissionCapability, &permission.Capability{
+	// always register the plus extended permission capability
+	_, err = plusManager.RegisterCapability(rportplus.PlusExtendedPermissionCapability, &extendedpermission.Capability{
 		Config: nil,
 		Logger: logger,
 	})
 	if err != nil {
-		return fmt.Errorf("unable to register plus permission capability: %w", err)
+		return fmt.Errorf("unable to register plus extended permission capability: %w", err)
 	}
-	logger.Infof("plus permission capability registered")
+	logger.Infof("plus extended permission capability registered")
 
 	return nil
 }
